Drop the consumer-less health latch: getHealth reads the registry stamps
healthState/HealthSignal were built expecting the read server would serve
getHealth through them; it serves through the registry's close-time stamps
instead, leaving the latch write-only. The loop stamps the registry directly.
